add test cases around org/no org, flow without org not cause issues
