+++ cumulusci/core/utils.py
@@ -1,7 +1,6 @@
 """ Utilities for CumulusCI Core
 
-import_class: Task class defn import helper
-MockLoggingHandler: a handler for testing logging """
+import_class: Task class defn import helper """
 
 import logging
 
@@ -13,33 +12,3 @@ def import_class(path):
     module = '.'.join(module)
     mod = __import__(module, fromlist=[components[-1]])
     return getattr(mod, components[-1])
-
-
-class MockLoggingHandler(logging.Handler):
-    """Mock logging handler to check for expected logs.
-
-    Messages are available from an instance's ``messages`` dict, in order, indexed by
-    a lowercase log level string (e.g., 'debug', 'info', etc.).
-    """
-
-    def __init__(self, *args, **kwargs):
-        self.messages = {'debug': [], 'info': [], 'warning': [], 'error': [],
-                         'critical': []}
-        super(MockLoggingHandler, self).__init__(*args, **kwargs)
-
-    def emit(self, record):
-        "Store a message from ``record`` in the instance's ``messages`` dict."
-        self.acquire()
-        try:
-            self.messages[record.levelname.lower()].append(record.getMessage())
-        finally:
-            self.release()
-
-    def reset(self):
-        """ Reset the handler in TestCase.setUp() to clear the msg list """
-        self.acquire()
-        try:
-            for message_list in self.messages.values():
-                del message_list[:]
-        finally:
-            self.release()
